Calculo de Isothermal Entropy Change
