+++ spinsim/thermodynamic/derivates.py
@@ -1,0 +1,24 @@
+from .compute_observables import *
+import numpy as np
+
+
+"""
+Calculo de derivadas con ciertas variables fijas
+"""
+
+"""
+Derivada magnetizacion respecto a la temperatura
+"""
+
+"""
+Derivada magnetizacion respecto al campo
+"""
+
+"""
+Derivada entropia respecto a la temperatura
+"""
+
+"""
+Derivada entropia respecto al campo
+"""
+
